Lay the Python 3 groundwork with an automated pass
And so begins the long process of modernizing the code. This is the first
major step which involved using an automated conversion tool. For this, the
fissix tool was used. Fissix, the modern successor to 2to3, swept through
the code making many small syntax corrections. Now the real work begins...