Build a better mousetrap
The old monolithic mouse module was a tangled nest of code, a clumsy trap that was
difficult for everyone. Following our strategy with the keyboard changes, we've tossed
out the old blueprints and built a better mousetrap.

The original module was deconstructed into a suite of specialized tools consisting of:

icewmcp_mouse_speed.py: A tool for mouse acceleration. The original had a 7 second
                        safety revert dialog which was ported as here as a clever
                        escape hatch if users accidentally set their speed too high.

icewmcp_mouse_cursors.py: This provides a clean UI for swapping cursor themes and
                          features image previews for each of the selected cursors.

icewmcp_mouse_behavior_stub.py: A consistent placeholder for functionality that will
                                eventually be untangled from `icepref` at a later date.